Disable charger daemon
When you plug charger when the phone is off, the phone will start
up with special flags in command line. On android, this will make
other services to stop and starting charger. As Ubuntu touch
doesn't understand this yet, this will also make Ubuntu touch
partially starts (as most of services in android container is
stoped). So, this commit disable charger daemon and it's init
config. Now, when you plug charger in, the phone will just
boot straight to Ubuntu touch.