Change the way to distribute Pulseaudio module.
Previously, we ship files directly in device tarball. The downside
of that method is our files can be overwritten if Ubuntu base (delta)
image includes update to Pulseaudio. In this commit, we changed to
including the files in Android base image. This way, our files will
get bind-mounted, and changes in Ubuntu base image will not affect
our files.

This is based on Pulseaudio version 6.0-0ubuntu9.8 patched with
custom patch. Get your patch here:
https://github.com/peat-psuwit/pulseaudio-droid-card-parameters-setter/