udev.rules: add the rule for video node's link.
These nodes are used for video encoding and decoding. By setting the
links this way, the library won't have to use the hard-coded node
number.

Actually, this behavior is coded in ueventd (in CM code, at least).
So, they should be in some upstream udev rules. For now, just put it
here.