Add a service to kickstart wlan module in kernel when wcnss_service is ready.
Prima WLAN driver in Linux, when compiled as a built-in module, have to
be noified when wcnss_service is ready. This commit add the service that
runs when wcnss_service runs. The problem is that when wcnss_service starts,
it doesn't necessarily be ready. So, the service will keep notifying module
until wlan0 interface appear on the system.